refactor: taskListFilter함수 TypeTimeTable 파일에서 분리
