+++ src/app/timetable/components/TypeTimeTable/index.tsx
@@ -1,7 +1,7 @@
 import { useContext } from 'react';
 import Slot from './Slot';
 import { Task } from '../Timetable.type';
-import { generateClassNameWithType, hasKey, insertKey } from '../../utils';
+import { generateClassNameWithType, hasKey, insertKey, taskListFilter } from '../../utils';
 import TypeContext from '../../TypeContext';
 import styles from './TypeTimeTable.module.scss';
 import CurrentTimeLine from '../CurrentTimeLine';
@@ -19,19 +19,6 @@ interface TypeTimeTableProps {
   taskSlotStyle?: React.CSSProperties;
   timeTableStyle?: React.CSSProperties;
 }
-
-const taskListFilter = (taskListInput: Task[], checkHour: number, slotTimeInput: number) =>
-  taskListInput.filter((task: Task) => {
-    const taskStartHour = task.startTime.getHours();
-    const taskEndHour = task.endTime.getHours();
-    const taskEndMinute = task.endTime.getMinutes();
-
-    return (
-      taskStartHour <= checkHour &&
-      taskEndHour >= checkHour &&
-      !(taskEndHour === checkHour && taskEndMinute === slotTimeInput % 60)
-    );
-  });
 
 function TypeTimeTable({
   timeSlots,
